change into to float

+++ General-Function-Codes/WeightGraph.py
@@ -32,7 +32,8 @@ print("Welcome to Weight Reduce Generator")
 print("It shows how much weight you will lose in coming days")
 print("If you lose 1 % of your weight per week")
 
-present_weight = int(input("Input present weight, default 84.5 : ") or "84.5")
+present_weight = float(
+    input("Input present weight, default 84.5 : ") or "84.5")
 today = datetime.date.today()
 
 print(today)
